OnDemandRaster now in its own file
