+++ blob/gdal_raster.h
@@ -65,6 +65,33 @@ namespace spacepop {
             }
         }
     }
+
+
+//! Convert a lat-long into a specific pixel.
+    std::array<int, 2> pixel_containing(std::array<double, 2> coord, const std::vector<double>& transform);
+
+//! Convert a pixel corner into a lat-long
+    template<typename POINTISH>
+    POINTISH pixel_coord(std::array<int, 2> pixel, const std::vector<double>& transform) {
+        return {
+                transform[0] + pixel[0] * transform[1] + pixel[1] * transform[2],
+                transform[3] + pixel[1] * transform[4] + pixel[5]
+        };
+    }
+
+
+//! Convert a pixel into its four corners as lat-long.
+//  Boost::polygon likes to be clockwise, so these are clockwise.
+    template<typename POINTISH>
+    std::vector<POINTISH> pixel_bounds(std::array<int, 2> pixel, const std::vector<double>& transform) {
+        return {
+                pixel_coord<POINTISH>(pixel, transform),
+                pixel_coord<POINTISH>({pixel[0], pixel[1] + 1}, transform),
+                pixel_coord<POINTISH>({pixel[0] + 1, pixel[1] + 1}, transform),
+                pixel_coord<POINTISH>({pixel[0] + 1, pixel[1]}, transform)
+        };
+    }
+
 }
 
 #endif //BLOB_GDAL_RASTER_H
